chore: pull CRS from `common` into `acvm-backend-barretenberg`
