+++ cspell.json
@@ -19,6 +19,7 @@
         "linearization",
         "logstr",
         "merkle",
+        "nargo",
         "PAIRINGSBN",
         "pedersen",
         "peekable",
